% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew.R
\docType{data}
\name{crew}
\alias{crew}
\title{The Crew.}
\format{A \code{sf} data frame with 4 rows and 3 variables:
\describe{
  \item{name}{nickname of the crew member}
  \item{url}{url to recognition video}
  \item{geometry}{coordinates of where to find the dude}
  }}
\usage{
crew
}
\description{
A dataset containing the crew member nicknames, their location and
their recognition videos.
}
\keyword{datasets}
