% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here_comes.R
\name{here_comes}
\alias{here_comes}
\title{Here comes the Hotstepper}
\usage{
here_comes(who = c("Hotstepper", "SpinDoctor", "KingJames", "Hackerman",
  "Dirtydan"))
}
\arguments{
\item{who}{\code{character}; at the moment one of:
\itemize{
\item HotStepper
\item SpinDoctor
\item KingJames
\item Hackerman
\item Dirtydan}}
}
\examples{
  here_comes("Hotstepper")

}
\author{
Christoph Stepper
}
