% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/here_lives.R
\name{here_lives}
\alias{here_lives}
\title{The Crew lives here}
\usage{
here_lives(crowd = crew, all = FALSE)
}
\arguments{
\item{crowd}{\code{sf} data frame. Check the package data \emph{crew}
to see the data structure.}

\item{all}{\code{logical}. Should all crew members be shown.
If \code{FALSE}, user input is evaluated to create subset.}
}
\value{
\code{mapview} object.
}
\examples{
  options(viewer = NULL)
  crew_map = here_lives(crew, all = TRUE)
  mapview(crew_map)

}
\author{
Christoph Stepper

Tim Appelhans
}
