% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herelives.R
\name{herelives}
\alias{herelives}
\title{The Crew lives here}
\usage{
herelives(crowd = crew, all = FALSE)
}
\arguments{
\item{crowd}{\code{sf} data frame. Check the package data \emph{crew}
to see the data structure.}

\item{all}{\code{logical}. Should all crew members be shown.
If \code{FALSE}, user input is evaluated to create subset.}
}
\value{
\code{mapview} object.
}
\examples{
  herelives(crew, all = TRUE)

}
\author{
Christoph Stepper

Tim Appelhans
}
