% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hottest_crew}
\alias{hottest_crew}
\title{The Hottest Crew}
\format{A \code{sf} data frame with 5 rows and 3 variables:
\describe{
  \item{name}{nickname of the crew member}
  \item{youtube}{youtube ID to recognition video}
  \item{geometry}{coordinates of where to find the dude}
  }}
\usage{
hottest_crew
}
\description{
A dataset containing the crew member nicknames, their location and
their recognition videos.
}
\keyword{datasets}
