% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_crew_member.R
\name{identify_crew_member}
\alias{identify_crew_member}
\title{Identify Crew Members}
\usage{
identify_crew_member(crew = hottest_crew, all_members = FALSE)
}
\arguments{
\item{crew}{\code{sf} data frame. Check the package data \emph{hottest_crew}
to see the data structure.}

\item{all_members}{\code{logical}. Should all crew members be selected.
If \code{FALSE}, user input is evaluated to create subset.}
}
\value{
\code{sf} data frame.
}
\author{
Christoph Stepper
}
